Project modules reorganisation
